+++ src/vs/workbench/contrib/accessibility/browser/accessibility.contribution.ts
@@ -26,11 +26,10 @@ import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
 import { getNotificationFromContext } from 'vs/workbench/browser/parts/notifications/notificationsCommands';
 import { IListService, WorkbenchList } from 'vs/platform/list/browser/listService';
 import { NotificationFocusedContext } from 'vs/workbench/common/contextkeys';
-import { IAccessibleViewService, AccessibleViewService, IAccessibleContentProvider, IAccessibleViewOptions, AccessibleViewType } from 'vs/workbench/contrib/accessibility/browser/accessibleView';
+import { IAccessibleViewService, AccessibleViewService, IAccessibleContentProvider, IAccessibleViewOptions, AccessibleViewType, AccessibilityHelpAction, AccessibleViewAction } from 'vs/workbench/contrib/accessibility/browser/accessibleView';
 import { IHoverService } from 'vs/workbench/services/hover/browser/hover';
 import { alert } from 'vs/base/browser/ui/aria/aria';
 import { UnfocusedViewDimmingContribution } from 'vs/workbench/contrib/accessibility/browser/unfocusedViewDimmingContribution';
-import { AccessibilityHelpAction, AccessibleViewAction, AccessibleViewNextAction, AccessibleViewPreviousAction } from 'vs/workbench/contrib/accessibility/browser/accessibilityMultiCommands';
 
 registerAccessibilityConfiguration();
 registerSingleton(IAccessibleViewService, AccessibleViewService, InstantiationType.Delayed);
@@ -239,25 +238,6 @@ class NotificationAccessibleViewContribution extends Disposable {
 }
 
 workbenchContributionsRegistry.registerWorkbenchContribution(NotificationAccessibleViewContribution, LifecyclePhase.Eventually);
-
-class AccessibleViewNavigatorContribution extends Disposable {
-	static ID: 'AccessibleViewNavigatorContribution';
-	constructor() {
-		super();
-		this._register(AccessibleViewNextAction.addImplementation(95, 'next', accessor => {
-			const accessibleViewService = accessor.get(IAccessibleViewService);
-			accessibleViewService.next();
-			return true;
-		}));
-		this._register(AccessibleViewPreviousAction.addImplementation(95, 'previous', accessor => {
-			const accessibleViewService = accessor.get(IAccessibleViewService);
-			accessibleViewService.previous();
-			return true;
-		}));
-	}
-}
-
-workbenchContributionsRegistry.registerWorkbenchContribution(AccessibleViewNavigatorContribution, LifecyclePhase.Eventually);
 
 export function alertFocusChange(index: number | undefined, length: number | undefined, type: 'next' | 'previous'): void {
 	if (index === undefined || length === undefined) {
