+++ src/vs/workbench/contrib/inlineChat/browser/inlineChatActions.ts
@@ -31,7 +31,7 @@ import { Disposable } from 'vs/base/common/lifecycle';
 import { CommandsRegistry } from 'vs/platform/commands/common/commands';
 import { Position } from 'vs/editor/common/core/position';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { AccessibilityHelpAction } from 'vs/workbench/contrib/accessibility/browser/accessibilityMultiCommands';
+import { AccessibilityHelpAction } from 'vs/workbench/contrib/accessibility/browser/accessibleView';
 
 CommandsRegistry.registerCommandAlias('interactiveEditor.start', 'inlineChat.start');
